MA-16832 Revert "MA-16457 don't reboot if no display connect in recovery"
This reverts commit 7be662294bfb7b8025178ec4267c747cc4c3bfdf.

In recovery, "ShowMenu" function is used to get the key operations,
based on the key operations, the UI may be refreshed or the system
will reboot.

If display is not connected, "ShowMenu" will return "REBOOT" directly,
the reverted patch ignores this "REBOOT", "ShowMenu" is invoked in
deadloop to return "REBOOT".

A patch from Google will be picked here, the patch has "ShowMenu" not
return direclty if the display is not connected and wait for the key
operation, and with that patch, when "fastboot reboot bootloader"
command is executed, the bootloader message can be written to misc
partition.

Change-Id: Iad3006d22dd37d99ef3cb77c82f52928bdbc84b8
Reviewed-on: http://androidsource.ap.freescale.net/project/12535
Reviewed-by: Wang Haoran <elven.wang@nxp.com>