MA-21767 build the gen_update_config.py in to a hermetic executable
In Android14, the gen_update_config.py depends on the module
"apex_manifest" which depends on a intermediate module in a python file
generated from apex_manifest.proto. It is not easy to specify the paths
of the modules when execute this python file, so build it as a hermetic
executable.

AbOtaPropertyFiles is defined in ota_utils.py. The module ota_utils is
imported into ota_from_target_files.py, but AbOtaPropertyFiles is not.
This causes execution errors with the following log when
gen_update_config.py is built into an executable.

    AttributeError: module 'ota_from_target_files' has no attribute 'AbOtaPropertyFiles'

Refer the AbOtaPropertyFiles from ota_from_target_files.ota_utils
instead.

Change-Id: Icc39b2c004eb909e6d75c40ab9791a2095d1639e
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/22981
Reviewed-by: Elven Wang <elven.wang@nxp.com>